fix(typings): Correct typings file path
The package.json file previously referred to the typings file,
aurelia-middleware-koa.d.ts, in the ./dist directory. The typings file
is not in the ./dist directory, it is generated in each of the child
directories for the various module types (e.g. amd, commonjs).

For consistency with other Aurelia projects, this commit changes the
typings field in the package.json file to reference:
./dist/commonjs/aurelia-middleware-koa.d.ts